+++ test/test_fcl_broadphase_collision_1.cpp
@@ -45,6 +45,7 @@
 #include "fcl/broadphase/broadphase_interval_tree.h"
 #include "fcl/broadphase/broadphase_dynamic_AABB_tree.h"
 #include "fcl/broadphase/broadphase_dynamic_AABB_tree_array.h"
+#include "fcl/broadphase/default_broadphase_callbacks.h"
 #include "fcl/broadphase/detail/sparse_hash_table.h"
 #include "fcl/broadphase/detail/spatial_hash.h"
 #include "fcl/geometry/geometric_shape_to_BVH_model.h"
@@ -447,7 +448,7 @@ void broad_phase_update_collision_test(S env_scale, std::size_t env_size, std::s
     ts[i].push_back(timers[i].getElapsedTime());
   }
 
-  std::vector<test::CollisionData<S>> self_data(managers.size());
+  std::vector<DefaultCollisionData<S>> self_data(managers.size());
   for(size_t i = 0; i < managers.size(); ++i)
   {
     if(exhaustive) self_data[i].request.num_max_contacts = 100000;
@@ -457,7 +458,7 @@ void broad_phase_update_collision_test(S env_scale, std::size_t env_size, std::s
   for(size_t i = 0; i < managers.size(); ++i)
   {
     timers[i].start();
-    managers[i]->collide(&self_data[i], test::defaultCollisionFunction);
+    managers[i]->collide(&self_data[i], DefaultCollisionFunction);
     timers[i].stop();
     ts[i].push_back(timers[i].getElapsedTime());
   }
@@ -488,7 +489,7 @@ void broad_phase_update_collision_test(S env_scale, std::size_t env_size, std::s
 
   for(size_t i = 0; i < query.size(); ++i)
   {
-    std::vector<test::CollisionData<S>> query_data(managers.size());
+    std::vector<DefaultCollisionData<S>> query_data(managers.size());
     for(size_t j = 0; j < query_data.size(); ++j)
     {
       if(exhaustive) query_data[j].request.num_max_contacts = 100000;
@@ -498,7 +499,7 @@ void broad_phase_update_collision_test(S env_scale, std::size_t env_size, std::s
     for(size_t j = 0; j < query_data.size(); ++j)
     {
       timers[j].start();
-      managers[j]->collide(query[i], &query_data[j], test::defaultCollisionFunction);
+      managers[j]->collide(query[i], &query_data[j], DefaultCollisionFunction);
       timers[j].stop();
       ts[j].push_back(timers[j].getElapsedTime());
     }